[AIRFLOW-5297] Move AdlsToGcs operator to core
For more information check AIP-21.
