+++ tests/operators/test_adls_to_gcs_operator.py
@@ -19,7 +19,7 @@
 
 import unittest
 
-from airflow.contrib.operators.adls_to_gcs import \
+from airflow.operators.adls_to_gcs import \
     AdlsToGoogleCloudStorageOperator
 from tests.compat import mock
 
@@ -52,10 +52,10 @@ class TestAdlsToGoogleCloudStorageOperator(unittest.TestCase):
         self.assertEqual(operator.gcp_conn_id, GCS_CONN_ID)
         self.assertEqual(operator.azure_data_lake_conn_id, AZURE_CONN_ID)
 
-    @mock.patch('airflow.contrib.operators.adls_to_gcs.AzureDataLakeHook')
+    @mock.patch('airflow.operators.adls_to_gcs.AzureDataLakeHook')
     @mock.patch('airflow.contrib.operators.adls_list_operator.AzureDataLakeHook')
     @mock.patch(
-        'airflow.contrib.operators.adls_to_gcs.GoogleCloudStorageHook')
+        'airflow.operators.adls_to_gcs.GoogleCloudStorageHook')
     def test_execute(self, gcs_mock_hook, adls_one_mock_hook, adls_two_mock_hook):
         """Test the execute function when the run is successful."""
 
@@ -101,10 +101,10 @@ class TestAdlsToGoogleCloudStorageOperator(unittest.TestCase):
         # we expect MOCK_FILES to be uploaded
         self.assertEqual(sorted(MOCK_FILES), sorted(uploaded_files))
 
-    @mock.patch('airflow.contrib.operators.adls_to_gcs.AzureDataLakeHook')
+    @mock.patch('airflow.operators.adls_to_gcs.AzureDataLakeHook')
     @mock.patch('airflow.contrib.operators.adls_list_operator.AzureDataLakeHook')
     @mock.patch(
-        'airflow.contrib.operators.adls_to_gcs.GoogleCloudStorageHook')
+        'airflow.operators.adls_to_gcs.GoogleCloudStorageHook')
     def test_execute_with_gzip(self, gcs_mock_hook, adls_one_mock_hook, adls_two_mock_hook):
         """Test the execute function when the run is successful."""
